+++ NEWS.md
@@ -1,5 +1,8 @@
 # phenoptrReports 0.2.4.9000
 
+Analysis addin:
+- Reverse the y-axis of heatmaps so that the identity runs 
+  from the top left to bottom right. 
 Spatial map viewer improvements:
 - Spatial map viewer works correctly with inForm data containing 
   field coordinates (vs slide coordinates) (#16).